refactor: useValidateInput,useValidateArrayInput으로 파일 이름 변경
