+++ hooks/src/lib/useValidateArrayInput/useValidateArrayInput.ts
